Fix tests related to entities without an id.
Initialise the SessionFactory in IdGenerationTest with the correct
packages. This necessitated the move of an "invalid" entity
(NeitherGraphIdNorId) in its own class and package, otherwise the SessionFactory
initialisation would fail because of the other invalid entities in the
original package.

`saveWithContextIdStrategy` test needs its own custom SessionFactory,
so instantiate one in the test directly and don't pollute the static
context.

This fixing of the tests leads to a failing build because now the tests
expose a bug in the production code.